Échelles d'avarice

Résultats d'un questionnaire auprès de 300 participants administré par Mechanical Turks. Les questions 12, 13 et 25 sont inversées par rapport aux autres questions. Les scores sont mesurés à l'aide d'échelles de Likert qui vont de fortement en désaccord (1) à fortement en accord (5).

Utilisation

avarice

Format

Une base de données avec 300 lignes et 32 variables:

q1

Q1: I always want more

q2

Q2. Actually, I'm kind of greedy

q3

Q3: One can never have too much money

q4

Q4. As soon as I have acquired something, I start to think about the next thing I want

q5

Q5. It doesn't matter how much I have, I'm never completely satisfied

q6

Q6. My life motto is 'more is better'

q7

Q7. I can't imagine having too many things

q8

Q8. No matter how much I have of something, I always want more

q9

Q9. One can never have enough

q10

Q10. Even when I am fulfilled, I often seek more

q11

Q11. The pursuit of more and better is an important goal in life for me

q12

Q12. A simple basic life is sufficient for me

q13

Q13. I am easily satisfied with what I have got

q14

Q14. When I think about all the things I have, my first thought is about what I would like to have next

q15

Q15. My actions are strongly focused on material things

q16

Q16. Sometimes I feel a real urge to possess something

q17

Q17. When something is being shared, I try to get as big a share as possible

q18

Q18. In order to get what I want, I can accept the fact that other people may suffer damage

q19

Q19. I get the most fun out of buying myself all sorts of things

q20

Q20. When I play on my own, I sometimes cheat a little

q21

Q21. I enjoy being a part of exclusive clubs or groups that are not open to everyone

q22

Q22. I do not enjoy sharing positions of power

q23

Q23. I like to collect expensive things

q24

Q24. At work/school, I keep good ideas to myself so that only I can get credit for them in the long run

q25

Q25. Financially supporting the less fortunate is a priority for me

q26

Q26. I believe that money is essential; friends are replaceable

q27

Q27. Being financially wealthy is my number one goal.

q28

Q28. I consider myself successful if I have a job that pays a lot of money

q29

Q29. No matter how much I have, I always want more

q30

Q30. "I want it all" would be a good motto for me

age

âge

genre

genre de la personne, soit homme ou femme

Source

ResearchBox 159, https://researchbox.org/159, licence CC-BY-NC 4.0

Références

Zeelenberg M, Seuntjens TG, van de Ven, N et S.M. Breugelmans (2021). Dispositional Greed Scales. European Journal of Psychological Assessment. 1-10, doi: 10.1027/1015-5759/a000647