Étude sur le chômage

Dans le cadre d’un programme gouvernemental, on s’intéresse au nombre de semaines de paiements écoulé entre une perte d’emploi et un nouvel emploi pour des chômeurs et chômeuses âgé(e)s entre 18 et 65 ans, dans le but d’élaborer un programme de formation destiné au chômage de longue durée. La base de données contient un échantillon fictif de 2 500 personnes inscrites à l’assurance emploi entre janvier et novembre 2021

Utilisation

chomage

Format

Une base de données avec 2500 lignes et 8 variables:

duree

double, nombre de semaines de prestations payées par le programme d’assurance emploi.

age

facteur, variable catégorielle pour la tranche d’âge, soit 18,25, 25,50 ou 50,65 ans.

formation

facteur, variable catégorielle indiquant le niveau de formation, soit aucune, secondaire, professionnel, collegial ou universitaire.

sexe

entier, variable indicatrice binaire ; 1 pour femmes, 0 pour hommes.

retrait

entier, 1 si la personne bénéficiaire se désinscrit du programme avant la fin de la période maximale de prestation parce qu’elle cesse de chercher un emploi, 2 si elle trouve un emploi, 0 sinon.

montant

double, montant hebdomadaire des prestations, allant jusqu’à 55 pourcent de la rémunération hebdomadaire avant perte d’emploi jusqu’à concurrence de 547$ par semaine.

nheures

entier, nombre d’heures cumulées à l’emploi avant d’être au chômage, soit un minimum de 14 semaines (420 heures) ou 22 semaines (700 heures) dépendamment du taux de chômage de la région et du nombre d’heures d’emploi assurables dans la période de référence.

prestationmax

double, durée maximale pour les prestations

Auteur

Léo Belzile