Élections générales québécoises

Résultat électoraux pour les élections de 2018 et 2022 par circonscription en format large (parti, année). Les informations sociodémographiques sont basées sur les données du Recensement 2021 de Statistique Canada.

Utilisation

electionsqc

Format

Une base de données avec 125 lignes et 21 variables:

circonscription

nom de la circonscription électorale en 2022

CAQ_2022

pourcentage des suffrages exprimés pour la Coalition Avenir Québec en 2022

PLQ_2022

pourcentage des suffrages exprimés pour le Parti Libéral du Québec en 2022

QS_2022

pourcentage des suffrages exprimés pour Québec Solidaire en 2022

PQ_2022

pourcentage des suffrages exprimés pour le Parti Québécois en 2022

PCQ_2022

pourcentage des suffrages exprimés pour le Parti Conservateur du Québec en 2022

CAQ_2018

pourcentage des suffrages exprimés pour la Coalition Avenir Québec en 2018

PLQ_2018

pourcentage des suffrages exprimés pour le Parti Libéral du Québec en 2018

QS_2018

pourcentage des suffrages exprimés pour Québec Solidaire en 2018

PQ_2018

pourcentage des suffrages exprimés pour le Parti Québécois en 2018

PCQ_2018

pourcentage des suffrages exprimés pour le Parti Conservateur du Québec en 2018

partici_2018

taux de participation (en pourcentage) en 2018

partici_2022

taux de participation (en pourcentage) en 2022

age_30_59

pourcentage de la population totale âgée de 30 à 59 ans

age_60p

pourcentage de la population totale âgée de 60 ans et plus

age_med

âge médian

taille_famille

taille moyenne de la famille

fra

pourcentage de la population totale de langue maternelle française

ang

pourcentage de la population totale de langue maternelle anglaise

rev_tot_med

revenu total (avant impôts) médian

faible_rev

pourcentage de la population de plus de 18 ans à faible revenu (population applicable)

Source

Données ouvertes extraites du site du Directeur général des élections du Québec, licence https://www.dgeq.org/licence.html.

Détails

Comprend des données ouvertes octroyées sous la licence d'utilisation des données ouvertes du directeur général des élections disponible à l'adresse Web dgeq.org. L'octroi de la licence n'implique aucune approbation par le directeur général des élections de l'utilisation des données ouvertes qui en est faite.