Étude dans un magasin

Pour les besoins d’une enquête, on a demandé à 200 consommateurs adultes de répondre aux questions suivantes par rapport à un certain type de magasin sur une échelle de Likert de 1 (pas important) à 5 (très important) en jugeant de l'importance de 12 items.

Utilisation

factor

Format

Une base de données avec 200 lignes et 12 variables:

x1

le magasin offre de bons prix tous les jours?

x2

le magasin accepte les cartes de crédit majeures (Visa, Mastercard)?

x3

le magasin offre des produits de qualité?

x4

les vendeurs connaissent bien les produits?

x5

il y ait des ventes spéciales régulièrement?

x6

les marques connues soient disponibles?

x7

le magasin ait sa propre carte de crédit?

x8

le service soit rapide?

x9

il y ait une vaste sélection de produits?

x10

le magasin accepte le paiement par carte de débit?

x11

le personnel soit courtois?

x12

le magasin ait en stock les produits annoncés?

Auteur·rice

Denis Larocque