Discrimination salariale dans un collège

Une étude a été effectuée dans un collège américain pour explorer la distribution des salaires et vérifier s'il existe des inégalités salariales entre hommes et femmes.

Utilisation

salaireprof

Format

Une base de données de 397 observations contenant les variables suivantes

annees

nombre d'années depuis l'obtention du doctorat.

service

nombre d'années de service.

salaire

salaire de professeurs pendant l'année académique 2008–2009 (en milliers de dollars USD).

echelon

échelon académique, soit adjoint (adjoint), aggrégé (aggrege) ou titulaire (titulaire).

domaine

variable catégorielle indiquant le champ d'expertise du professeur, soit appliqué (applique) ou théorique (theoretique).

sexe

indicateur binaire pour le sexe, homme ou femme.

Source

Fox J. and Weisberg, S. (2019) An R Companion to Applied Regression, Third Edition, Sage.

Voir également