Sondage de satisfaction en entreprise

Une compagnie fait un sondage auprès de ses employés. Les questions Q8–Q17 retenues pour l’analyse factorielle dans la base de données sont mesurées à l’aide d’échelles de Likert allant de 1 (fortement en désaccord) à 5 (fortement d’accord).

Utilisation

sondage_entreprise

Format

Une base de données avec 482 lignes et 10 variables:

q8

La compagnie se préoccupe de la société

q9

La compagnie est dotée d’un grand sens éthique

q10

La compagnie est un meneur dans l’industrie des TI

q11

La compagnie est innovante

q12

La compagnie a des dirigeants forts, expérimentés et compétent

q13

La compagnie a une base financière saine

q14

Je peux faire confiance à la compagnie

q15

La compagnie a des publicités que j’apprécie particulièrement

q16

La compagnie est reconnue pour le bon traitement de ses employés

q17

La compagnie fait sa part pour aider la société

Détails

Source inconnue, données fournies par Julie Meloche