Analyse factorielle avec composantes principales

Cette fonction calcule une décomposition spectrale d'une matrice de corrélation ou de covariance des données et extrait les k premières composantes selon nfact, soit un entier ou le critère de Kaiser.

Utilisation

factocp(x, covmat = NULL, cor = TRUE, nfact = "kaiser")

Arguments

x

une matrice ou base de données

covmat

matrice de covariance

cor

logique; si TRUE, calculer la décomposition à partir de la matrice de corrélation plutôt que la matrice de covariance

nfact

entier pour le nombre de facteurs, ou chaîne de caractère "kaiser" pour le critère des valeurs propres.

Valeur de retour

une liste avec composante loadings contenant les chargements et si le nombre de facteurs est supérieur à un, la matrice de rotation varimax.

Exemples

factocp(covmat = cov(factor), nfact = 2)
#> $loadings
#> 
#> Loadings:
#>       F1     F2    
#>  [1,] -0.346       
#>  [2,]  0.281  0.128
#>  [3,]  0.156  0.779
#>  [4,]  0.731       
#>  [5,] -0.225       
#>  [6,]         0.660
#>  [7,]  0.405       
#>  [8,]  0.784 -0.104
#>  [9,]         0.737
#> [10,]  0.457       
#> [11,]  0.673       
#> [12,]         0.718
#> 
#>                   F1    F2
#> SS loadings    2.263 2.163
#> Proportion Var 0.189 0.180
#> Cumulative Var 0.189 0.369
#> 
#> $rotmat
#>            [,1]      [,2]
#> [1,]  0.7853218 0.6190878
#> [2,] -0.6190878 0.7853218
#> 
factocp(factor, nfact = "kaiser", cor = FALSE)
#> $loadings
#> 
#> Loadings:
#>     F1     F2     F3     F4     F5    
#> x1          0.109  0.129  0.883  0.281
#> x2  -0.106 -0.981                0.204
#> x3                -0.995 -0.127       
#> x4   1.118                       0.349
#> x5                        1.009 -0.267
#> x6                -0.561         0.705
#> x7         -0.809                     
#> x8   1.122 -0.153                     
#> x9                -0.882              
#> x10  0.132 -1.114 -0.111        -0.241
#> x11  1.062  0.157 -0.169 -0.120 -0.413
#> x12               -0.781         0.376
#> 
#>                   F1    F2    F3    F4    F5
#> SS loadings    3.695 2.930 2.758 1.845 1.188
#> Proportion Var 0.308 0.244 0.230 0.154 0.099
#> Cumulative Var 0.308 0.552 0.782 0.936 1.035
#> 
#> $rotmat
#>            [,1]        [,2]       [,3]        [,4]        [,5]
#> [1,] 0.90161230 -0.25071582 -0.2987722 -0.18256899  0.04050381
#> [2,] 0.40028783  0.75379412  0.4624123  0.14189682 -0.19391818
#> [3,] 0.09730667 -0.57630414  0.7796863 -0.05245829 -0.21850014
#> [4,] 0.11946882 -0.18604744 -0.0657347  0.97002381  0.07646121
#> [5,] 0.05588848  0.04686028  0.2909953 -0.05325221  0.95245291
#>